requirements.txt
opencv-python
mediapipe
Pillow
google-generativeai
cvzone
numpy
pynput
pyautogui
pycaw